+++ src/fields/formfields/Hidden.php
@@ -220,6 +220,7 @@ class Hidden extends FormField implements PreviewableFieldInterface
             SchemaHelper::cssClasses(),
             SchemaHelper::containerAttributesField(),
             SchemaHelper::inputAttributesField(),
+            SchemaHelper::enableContentEncryptionField(),
         ];
     }
 }