+++ src/fields/formfields/MultiLineText.php
@@ -254,6 +254,7 @@ class MultiLineText extends FormField implements PreviewableFieldInterface
             SchemaHelper::cssClasses(),
             SchemaHelper::containerAttributesField(),
             SchemaHelper::inputAttributesField(),
+            SchemaHelper::enableContentEncryptionField(),
         ];
     }