+++ src/fields/formfields/SingleLineText.php
@@ -270,6 +270,7 @@ class SingleLineText extends FormField implements PreviewableFieldInterface
             SchemaHelper::cssClasses(),
             SchemaHelper::containerAttributesField(),
             SchemaHelper::inputAttributesField(),
+            SchemaHelper::enableContentEncryptionField(),
         ];
     }