+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - Added page button conditions, to conditionally show/hide next button according to your logic.
 - Added page conditions, to conditionally show/hide page according to your logic.
 - All-new page settings manager for form builder.
+- Added min and max date options for Date fields.
+- Added "Enable Content Encryption" setting on Address, Email, Hidden, Multi-Line Text, Name, Phone, Recipients and Single-Line Text fields. This will encrypt submission content in the database, preventing human-readable data for sensitive fields.
+
+- Date fields can now use [Flatpickr.js]() when rendered as a calendar.
+- Date fields can now content-manage their date and time format.
 
 - Fixed JS errors showing in form builder error alert.
 - Fixed page errors not showing on page labels in the control panel form builder.